criado a clase de dto e controllers, feito testes com o insomminia e incuido a dependencia de validação

+++ back_end/Gestao_Escolar/src/main/java/com/example/GestaoEscolar/models/CadastroModel.java
@@ -4,10 +4,9 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
+import lombok.*;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Data
@@ -19,7 +18,7 @@ public class CadastroModel {
     @GeneratedValue(strategy = GenerationType.UUID)
     private UUID id;
     private String nomeCompleto;
-    private Integer ra;
+    private BigDecimal ra;
     private String endereco;
     private String bairro;
     private Integer cep;
